ISSUE #2345 - fix errors
